Move DropwizardSSLConnectionSocketFactoryTest to io.dropwizard.client (#3657)
This puts it in the same package as the class it's testing, which lets us expose the HostnameVerifier to the tests without relying on reflection of a private field.
